requirements.txt
streamlit==1.24.0
transformers==4.31.0
torch==2.0.1
pytesseract==0.3.10
opencv-python-headless==4.7.0.72
Pillow==9.5.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
numpy==1.24.3
scikit-image==0.21.0
pytest==7.4.0
python-dotenv==1.0.0
